style: modify the YML file to double quotation marks

+++ packages/basic/index.js
@@ -327,7 +327,7 @@ module.exports = {
     ],
 
     // yml
-    "yml/quotes": ["error", { prefer: "single", avoidEscape: false }],
+    "yml/quotes": ["error", { prefer: "double", avoidEscape: false }],
     "yml/no-empty-document": "off",
 
     // kirklin
